feat: 애플 로그인 구현 테스트
